pyproject.toml
[project]
name = "clawsmith"
version = "0.1.0"
description = "AI-powered code orchestration pipeline with intelligent model routing"
requires-python = ">=3.11"
license = {file = "LICENSE"}
dependencies = [
    "mcp[cli]>=1.7.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "click>=8.1",
    "httpx>=0.27",
    "litellm>=1.40",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
clawsmith = "orchestrator.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[tool.setuptools.packages.find]
where = ["."]
include = ["orchestrator*", "mcp_server*", "providers*", "routing*", "tools*", "jobs*", "config*", "prompts*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
